backend/TutorAI_v3/backend_requirements.txt
# Core Django backend
Django==4.2.25
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.4

# Numeric + science
numpy==1.21.6
scipy==1.10.1

# ML / audio deps used in views.py
scikit-learn==1.3.2
openai-whisper
TTS==0.14.3